+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add delayMs property to workflow management for enhanced playback control. (`eff9c395`)
 - Enhance workflow execution and metadata handling. (`1390072c`)
 - Enhance workflow action management with UUID integration. (`9647b650`)
 - Enhance workflow management and plugin integration. (`e55d2821`)